requirements.txt
django==4.2.7
django-extensions
ipython
django-debug-toolbar
django-admin-rangefilter
django-import-export

#flake8
flake8
flake8-builtins
flake8-print
flake8-breakpoint
